requirements.txt
fastapi
pydantic
pydantic_settings
uvicorn
transformers
safetensors
torch
psycopg2-binary